just moved around a class
